fixes the lint compliance issues
