+++ src/main/java/metrics/LightChainCollector.java
@@ -1,19 +1,22 @@
 package metrics;
 
-public class LightChainCollector implements Collector{
+/**
+ * Collector of metrics for the LightChain implementation of Prometheus equivalents.
+ */
+public class LightChainCollector implements Collector {
 
-    @Override
-    public HistogramCollector histogram() {
-        return new LightChainHistogram();
-    }
+  @Override
+  public HistogramCollector histogram() {
+    return new LightChainHistogram();
+  }
 
-    @Override
-    public GaugeCollector gauge() {
-        return new LightChainGauge();
-    }
+  @Override
+  public GaugeCollector gauge() {
+    return new LightChainGauge();
+  }
 
-    @Override
-    public CounterCollector counter() {
-        return new LightChainCounter();
-    }
+  @Override
+  public CounterCollector counter() {
+    return new LightChainCounter();
+  }
 }
